django/pyproject.toml
[project]
name = "cicd-testing"
version = "0.1.0"
description = "CI/CD Testing: testing django cicd"
authors = ["RRCHNM <chnm@gmu.edu>"]
license = "WTFPL"
readme = "README.md"
requires-python = ">=3.13"
packages = [{include = "cicd_testing"}]

dependencies = [
  #"python >= 3.13",
  "python-dotenv ~= 1.0.1",
  "daphne ~= 4.2.1",
  "whitenoise ~= 6.11.0",
  "django ~= 5.2.7",
  "django-environ ~= 0.12.0",
  # allauth
  "requests",
  "django-allauth ~= 65.13.0",
  
  "django-tailwind ~= 4.3.1",
  
  # postgres
  "psycopg2-binary ~= 2.9.11",

  # obj storage
  "django-storages ~= 1.14.6",
  "boto3 ~= 1.40.64",
]

[dependency-groups]
dev = [
  "black ~= 24.10.0",
  "djhtml ~= 3.0.7",
  "pre-commit ~= 4.0.1",
  "django-debug-toolbar >= 6.1.0",
  "pytest-django ~= 4.11.1",
  "pylint-django ~= 2.6.1",
]
